comamnds: Break into functional groups
Split main commands.js into a directory module, where commands
are organized into separate files by functional group. A single
index.js handles registration of all commands into a central list.

A simple unit test ensures that all commands have the right
basic "signature".

Without any dependence on Discord, we can now create `client.commands`
directly during app start up instead of waiting for connection to
the server. This removes the need for `loadCommands`.
