+++ commands/helpers.js
@@ -1,0 +1,44 @@
+const settings = require('../settings/settings.json')
+
+const hd = require('humanize-duration')
+
+function requireRole (member, roleName = 'Bot Manager', errorMessage = 'You require the bot manager role to use this command.') {
+  if (!member.roles.find(r => r.name === roleName) || !settings.pinano_guilds.includes(member.guild.id)) {
+    throw new Error(errorMessage)
+  }
+}
+
+function requireParameterCount (args, argCount, usageStr) {
+  if (args.length !== argCount) {
+    throw new Error(`Usage: \`${usageStr}\``)
+  }
+}
+
+function requireParameterFormat (arg, formatFn, usageStr) {
+  if (!formatFn(arg)) {
+    throw new Error(`Usage: \`${usageStr}\``)
+  }
+}
+
+async function selfDestructMessage (messageFn) {
+  let m = await messageFn()
+  setTimeout(() => m.delete(), settings.res_destruct_time * 1000)
+}
+
+function abbreviateTime (playtime) {
+  return hd(playtime * 1000, { units: ['h', 'm', 's'], round: true })
+    .replace('hours', 'h')
+    .replace('minutes', 'm')
+    .replace('seconds', 's')
+    .replace('hour', 'h')
+    .replace('minute', 'm')
+    .replace('second', 's')
+}
+
+module.exports = {
+  requireRole,
+  requireParameterCount,
+  requireParameterFormat,
+  selfDestructMessage,
+  abbreviateTime
+}
